# docs/full-delivery/openapi.json
{
  "components": {
    "schemas": {
...
      "AddRepository": {
        "additionalProperties": false,
        "properties": {
          "ref": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref"
          },
          "repository": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Repository",
            "type": "string"
          }
        },
        "required": [
          "repository"
        ],
        "title": "AddRepository",
        "type": "object"
      },
...
      "DocumentRecord": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "code_license": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicenseObservation"
              },
              {
                "type": "null"
              }
            ]
          },
          "code_location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CodeLocation"
              },
              {
                "type": "null"
              }
            ]
          },
          "contract_version": {
            "const": "full-delivery-v1",
            "default": "full-delivery-v1",
            "title": "Contract Version",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Document Id",
            "type": "string"
          },
          "file_name": {
            "maxLength": 240,
            "minLength": 1,
            "title": "File Name",
            "type": "string"
          },
          "import_status": {
            "enum": [
              "READY",
              "PARTIAL",
              "NO_EXTRACTABLE_TEXT",
              "FAILED"
            ],
            "title": "Import Status",
            "type": "string"
          },
          "mode": {
            "enum": [
              "LIVE",
              "FIXTURE",
              "UNAVAILABLE"
            ],
            "title": "Mode",
            "type": "string"
          },
          "original_sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Original Sha256",
            "type": "string"
          },
          "revision": {
            "default": 1,
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "source_type": {
            "enum": [
              "PDF",
              "PPTX",
              "DOCX",
              "MARKDOWN",
              "CODE"
            ],
            "title": "Source Type",
            "type": "string"
          },
          "unit_count": {
            "minimum": 0.0,
            "title": "Unit Count",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "mode",
          "document_id",
          "file_name",
          "source_type",
          "original_sha256",
          "unit_count",
          "import_status",
          "created_at"
        ],
        "title": "DocumentRecord",
        "type": "object"
      },
...
      "LearningErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "contract_version": {
            "const": "full-delivery-v1",
            "default": "full-delivery-v1",
            "title": "Contract Version",
            "type": "string"
          },
          "needed_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Needed Action"
          },
          "request_id": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "user_message": {
            "title": "User Message",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "stage",
          "code",
          "user_message",
          "retryable"
        ],
        "title": "LearningErrorResponse",
        "type": "object"
      },
...
      "OpenFile": {
        "additionalProperties": false,
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentRecord"
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/DocumentUnit"
            },
            "title": "Units",
            "type": "array"
          }
        },
        "required": [
          "document",
          "units"
        ],
        "title": "OpenFile",
        "type": "object"
      },
...
      "Repository": {
        "additionalProperties": false,
        "properties": {
          "commit_sha": {
            "pattern": "^[a-f0-9]{40}$",
            "title": "Commit Sha",
            "type": "string"
          },
          "complete_tree": {
            "title": "Complete Tree",
            "type": "boolean"
          },
          "contract_version": {
            "const": "full-delivery-v1",
            "default": "full-delivery-v1",
            "title": "Contract Version",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "mode": {
            "enum": [
              "LIVE",
              "FIXTURE",
              "UNAVAILABLE"
            ],
            "title": "Mode",
            "type": "string"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "repository": {
            "maxLength": 200,
            "pattern": "^[A-Za-z0-9_.-]+/[A-Za-z0-9_.-]+$",
            "title": "Repository",
            "type": "string"
          },
          "repository_id": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Repository Id",
            "type": "string"
          },
          "revision": {
            "default": 1,
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "tree_sha": {
            "pattern": "^[a-f0-9]{40}$",
            "title": "Tree Sha",
            "type": "string"
          }
        },
        "required": [
          "mode",
          "repository_id",
          "repository",
          "ref",
          "commit_sha",
          "tree_sha",
          "complete_tree",
          "created_at"
        ],
        "title": "Repository",
        "type": "object"
      },
...
      "TreeEntry": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "folder",
              "file",
              "symlink",
              "submodule"
            ],
            "title": "Kind",
            "type": "string"
          },
          "path": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          },
          "sha": {
            "pattern": "^[a-f0-9]{40}$",
            "title": "Sha",
            "type": "string"
          },
          "size": {
            "default": 0,
            "minimum": 0.0,
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "path",
          "sha",
          "kind"
        ],
        "title": "TreeEntry",
        "type": "object"
      },
      "TreePage": {
        "additionalProperties": false,
        "properties": {
          "complete_tree": {
            "title": "Complete Tree",
            "type": "boolean"
          },
          "directory": {
            "title": "Directory",
            "type": "string"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/TreeEntry"
            },
            "title": "Entries",
            "type": "array"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "entries",
          "directory",
          "total",
          "offset",
          "complete_tree"
        ],
        "title": "TreePage",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Concept-to-Code Learning · capability-aware",
    "version": "0.2.0.dev0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/learning/v1/repositories": {
      "get": {
        "operationId": "repositories_api_learning_v1_repositories_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Repository"
                  },
                  "title": "Response Repositories Api Learning V1 Repositories Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Repositories",
        "tags": [
          "Full delivery v1",
          "Repository workspace"
        ]
      },
      "post": {
        "operationId": "add_api_learning_v1_repositories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRepository"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Repository"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Add",
        "tags": [
          "Full delivery v1",
          "Repository workspace"
        ]
      }
    },
    "/api/learning/v1/repositories/search": {
      "get": {
        "operationId": "search_api_learning_v1_repositories_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Search",
        "tags": [
          "Full delivery v1",
          "Repository workspace"
        ]
      }
    },
    "/api/learning/v1/repositories/{repository_id}": {
      "get": {
        "operationId": "repository_api_learning_v1_repositories__repository_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_.:-]+$",
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Repository"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Repository",
        "tags": [
          "Full delivery v1",
          "Repository workspace"
        ]
      }
    },
    "/api/learning/v1/repositories/{repository_id}/files": {
      "post": {
        "operationId": "open_file_api_learning_v1_repositories__repository_id__files_post",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_.:-]+$",
              "title": "Repository Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "maxLength": 1000,
              "minLength": 1,
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenFile"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Open File",
        "tags": [
          "Full delivery v1",
          "Repository workspace"
        ]
      }
    },
    "/api/learning/v1/repositories/{repository_id}/tree": {
      "get": {
        "operationId": "tree_api_learning_v1_repositories__repository_id__tree_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_.:-]+$",
              "title": "Repository Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "directory",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 1000,
              "title": "Directory",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 100,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 300,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Tree",
        "tags": [
          "Full delivery v1",
          "Repository workspace"
        ]
      }
    },
...
  }
}